siren and clock sounds
in the sound assets folder in the fmod file I have added events containing the processes of the clock ticking and the siren alarm.  When implementing them into the game I will need to try either or to see which one we like better in the game. or mix both.